hasim2col and re-running setup-paths errors

+++ scripts/setup-paths.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
-mkdir configs/ 2>/dev/null
-ln -s $PWD/src/main/scala/gemmini/GemminiCustomConfigs.scala configs/GemminiCustomConfigs.scala
+if [ ! -d configs ]; then
+    mkdir configs/
+fi
+
+if [ ! -f configs/GemminiCustomConfigs.scala ]; then
+    ln -s $PWD/src/main/scala/gemmini/GemminiCustomConfigs.scala configs/GemminiCustomConfigs.scala
+fi
